+++ changelog.md
@@ -39,7 +39,7 @@
 - Add Capacitor to ensure the app does not hang if an upload is `await`ed out of order, or is never consumed.
 - Add `errorHandler` option to allow customization of new UploadBuffer streams.
 - Improved documentation.
-- **BREAKING** Update `processRequest` to return `Promise<{body: *, finished: Promise<typeof undefined>}>`
+- **BREAKING** Update `processRequest` to return `Promise<{operations: *, finished: Promise<typeof undefined>}>`
 
 ## 5.0.0